+++ antarest/output/adapters.py
@@ -19,7 +19,7 @@ from typing_extensions import override
 
 from antarest.core.model import StudyPermissionType
 from antarest.output.service import IStudyMetadataProvider, StudyMetadata
-from antarest.output.storage.file.storage import FileStudyOutputs, IFileOutputsProvider
+from antarest.output.storage.file.abstract_storage import FileStudyOutputs, IFileOutputsProvider
 from antarest.study.model import DEFAULT_WORKSPACE_NAME
 from antarest.study.service import StudyService
 from antarest.study.storage.utils import assert_permission
