+++ antarest/output/storage/output_storage.py
@@ -36,6 +36,7 @@ logger = logging.getLogger(__name__)
 class OutputStorageType(StrEnum):
     IN_STUDY_FILE_TREE = "IN_STUDY_FILE_TREE"
     V2 = "V2"
+    OUTSIDE_STUDY_FILE_TREE = "OUTSIDE_STUDY_FILE_TREE"
 
 
 @dataclass(frozen=True)
